requirements.txt
tensorflow>=2.8.0
streamlit>=1.20.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
plotly>=5.0.0
pandas>=1.3.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
pillow>=8.3.0
scipy>=1.7.0
scikit-image>=0.18.0
requests>=2.25.0
pydantic>=1.8.0
python-multipart>=0.0.5
matplotlib>=3.5.0
seaborn>=0.11.0
